feat: export dialog panel
