+++ docs/demo/pure.md
@@ -1,0 +1,3 @@
+## pure-debug
+
+<code src="../examples/pure.tsx">
